whisper-service/requirements.txt
flask==3.0.3
openai-whisper==20240930
torch==2.3.1
